Working display change, working on countdown

+++ src/c/teams.h
@@ -19,7 +19,7 @@ typedef struct {
   int16_t vs_id; // index into TEAMS[]; -1 means no opponent (bye week, or opponent not in TEAMS[])
   uint8_t score;
   uint8_t vs_score;
-  uint32_t gametime;
+  unsigned long gametime;
   uint8_t ranking;
   uint8_t wins;
   uint8_t postseasonGames;
